+++ app/src/main/java/com/courier/example/ExampleService.kt
@@ -1,13 +1,8 @@
 package com.courier.example
 
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.content.Context
-import android.media.RingtoneManager
-import android.os.Build
-import androidx.core.app.NotificationCompat
 import com.courier.android.Courier
-import com.courier.android.notifications.CourierIntent
+import com.courier.android.notifications.CourierPushNotificationIntent
+import com.courier.android.notifications.presentNotification
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
 
@@ -15,50 +10,34 @@ class ExampleService: FirebaseMessagingService() {
 
     override fun onMessageReceived(message: RemoteMessage) {
         super.onMessageReceived(message)
+
+        // Notify the Courier SDK that a push was delivered
         Courier.onMessageReceived(message)
-        message.presentNotification(this, MainActivity::class.java)
+
+        // Create the PendingIntent that runs when the user taps the notification
+        // This intent targets your Activity and carries the original message payload
+        val intent = CourierPushNotificationIntent(
+            context = this,
+            target = MainActivity::class.java,
+            payload = message
+        )
+
+        // Show the notification to the user.
+        // Prefer data-only FCM so this service runs even in background/killed state.
+        // Fall back to notification fields if data keys are missing.
+        intent.presentNotification(
+            title = message.data["title"] ?: message.notification?.title,
+            body = message.data["body"] ?: message.notification?.body,
+        )
+
     }
 
     override fun onNewToken(token: String) {
         super.onNewToken(token)
+
+        // Register/refresh this device’s FCM token with Courier.
+        // The SDK caches and updates the token automatically and links it to the current user.
         Courier.onNewToken(token)
-    }
-
-}
-
-fun RemoteMessage.presentNotification(context: Context, handlingClass: Class<*>?, icon: Int = android.R.drawable.ic_dialog_info, settingsTitle: String = "Notification settings") {
-
-    try {
-
-        val channelId = "default"
-        val pendingIntent = CourierIntent(context, handlingClass, this).pendingIntent
-        val defaultSoundUri = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION)
-
-        val title = data["title"] ?: notification?.title ?: "Empty Title"
-        val body = data["body"] ?: notification?.body ?: "Empty Body"
-
-        val notificationBuilder = NotificationCompat.Builder(context, channelId)
-            .setSmallIcon(icon)
-            .setContentTitle(title)
-            .setContentText(body)
-            .setAutoCancel(true)
-            .setSound(defaultSoundUri)
-            .setPriority(NotificationCompat.PRIORITY_MAX)
-            .setContentIntent(pendingIntent)
-
-        val notificationManager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            val channel = NotificationChannel(channelId, settingsTitle, NotificationManager.IMPORTANCE_HIGH)
-            notificationManager.createNotificationChannel(channel)
-        }
-
-        val uuid = System.currentTimeMillis().toInt()
-        notificationManager.notify(uuid, notificationBuilder.build())
-
-    } catch (e: Exception) {
-
-        print(e)
 
     }
